delete unused people functionality
git-svn-id: https://plugins.svn.wordpress.org/events-made-easy/trunk@923178 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ eme_people.php
@@ -21,6 +21,9 @@ function eme_people_page() {
                $message = __("Person info has been updated.","eme");
       else
                $message = __("Couldn't update the person. Please try again.","eme");
+   } elseif (isset($_POST['eme_admin_action']) && $_POST['eme_admin_action'] == 'deleteunusedpeople') {
+      eme_delete_persons_without_bookings();
+      $message = __("People without bookings have been deleted.","eme");
    } elseif (isset ($_POST['persons']) && isset($_POST['eme_admin_action']) && $_POST['eme_admin_action'] == 'delete_people') {
       eme_people_delete_layout();
       // don't show the people table in this case, so we return from the function
@@ -139,15 +142,23 @@ function eme_people_delete_layout($message = "") {
       <fieldset><p><legend><?php _e('What should be done with the bookings done by these people?','eme'); ?></legend></p>
       <ul style="list-style:none;">
       <li><label><input type="radio" id="delete_option0" name="delete_option" value="delete_assoc_bookings" /><?php _e('Delete associated bookings','eme');?></label></li>
-      <li><input type="radio" id="delete_option1" name="delete_option" value="transfer_assoc_bookings" />
-      <label for="delete_option1"><?php _e('Transfer associated bookings to:','eme');?></label> <select name='to_person_id' id='to_person_id' class=''>
       <?php
       $other_persons=eme_get_persons("",$persons);
-      foreach ($other_persons as $person) {
-         print "<option value='".$person['person_id']."'>".eme_sanitize_html($person['person_name'])."</option>";
+      if ($other_persons) {
+         ?>
+         <li><input type="radio" id="delete_option1" name="delete_option" value="transfer_assoc_bookings" />
+         <label for="delete_option1"><?php _e('Transfer associated bookings to:','eme');?></label>
+         <select name='to_person_id' id='to_person_id' class=''>
+         <?php
+         foreach ($other_persons as $person) {
+            print "<option value='".$person['person_id']."'>".eme_sanitize_html($person['person_name'])."</option>";
+         }
+         ?>
+         </select>
+         <?php
       }
       ?>
-      </select></li>
+      </li>
       </ul></fieldset>
       <input type="hidden" name="eme_admin_action" value="confirm_delete_people" />
       <p class="submit"><input type="submit" name="submit" id="submit" class="button" value="Confirm Deletion"  /></p></div>
@@ -444,17 +455,25 @@ function eme_people_table($message="") {
 
    $persons = eme_get_persons();
    $destination = admin_url("admin.php?page=eme-people");
+   if($message != "") {
+      ?>
+      <div id='message' class='updated fade below-h2' style='background-color: rgb(255, 251, 204);'>
+         <p><?php print $message; ?></p>
+      </div>
+      <?php
+   }
+
    if (count($persons) < 1 ) {
       _e("No people have responded to your events yet!", 'eme');
    } else { 
    ?>
       <p><?php _e('This table shows the data about the people who responded to your events', 'eme') ?> </p> 
-      <?php if($message != "") { ?>
-            <div id='message' class='updated fade below-h2' style='background-color: rgb(255, 251, 204);'>
-               <p><?php print $message; ?></p>
-            </div>
-      <?php } ?>
-
+      <form id='people-deleteunused' method='post' action='<?php print $destination; ?>'>
+      <input type="hidden" name="eme_admin_action" value="deleteunusedpeople" />
+      <input type="submit" value="<?php _e ( 'Delete people without bookings','eme' ); ?>" name="doaction" id="doaction" class="button-primary action" />
+      </form>
+      <div class="clear"></div>
+      <br />
       <form id='people-filter' method='post' action='<?php print $destination; ?>'>
       <select name="eme_admin_action">
       <option value="-1" selected="selected"><?php _e ( 'Bulk Actions' ); ?></option>
@@ -571,14 +590,28 @@ function eme_get_person_id_by_wp_id($wp_id) {
    global $wpdb; 
    $people_table = $wpdb->prefix.PEOPLE_TBNAME;
    $wp_id = eme_sanitize_request($wp_id);
-   $sql = "SELECT person_id FROM $people_table WHERE wp_id = '$wp_id';" ;
+   $sql = "SELECT person_id FROM $people_table WHERE wp_id = '$wp_id'";
    return($wpdb->get_var($sql));
 }
 
 function eme_delete_person($person_id) {
    global $wpdb; 
    $people_table = $wpdb->prefix.PEOPLE_TBNAME;
-   $sql = "DELETE FROM $people_table WHERE person_id = '$person_id';" ;
+   $sql = "DELETE FROM $people_table WHERE person_id = '$person_id'";
+   $wpdb->query($sql);
+   return 1;
+}
+
+function eme_delete_persons_without_bookings() {
+   global $wpdb; 
+   $people_table = $wpdb->prefix.PEOPLE_TBNAME;
+   $bookings_table = $wpdb->prefix.BOOKINGS_TBNAME;
+   $events_table = $wpdb->prefix.EVENTS_TBNAME;
+   // first, clean up unreferenced bookings
+   $sql = "DELETE FROM $bookings_table WHERE event_id NOT IN (SELECT DISTINCT event_id FROM $events_table)";
+   $wpdb->query($sql);
+   // now, delete unreferenced persons
+   $sql = "DELETE FROM $people_table WHERE person_id NOT IN (SELECT DISTINCT person_id FROM $bookings_table)";
    $wpdb->query($sql);
    return 1;
 }
